@@ -1,1 +1,1 @@
-ruby-2.6.2
+ruby-2.7.2
