@@ -9,8 +9,7 @@ module Asyncapi
 
         def index
           jobs = paginate Job.all
-          # serializer = ActiveModel::ArraySerializer.new(jobs)
-          serializer = ActiveModel::Serializer::CollectionSerializer.new(jobs)
+          serializer = ActiveModel::ArraySerializer.new(jobs)
           render json: serializer
         end
 
